Fix function calls on function expressions stored in lists
Previously, attempting to call a function stored in a list (e.g., funcs[0](x))
resulted in "Function not found" because the interpreter did not properly
resolve function expressions inside IndexerExpressions.

This commit adds support for evaluating IndexerExpressions when calling functions,
allowing function expressions stored in lists to be correctly retrieved and executed.